Moved code for formatting calls
line_profiler/line_profiler_utils.py::CallbackRepr
    Migrated from `line_profiler/cleanup.py::_CallbackRepr`
